Add files via upload
This study is focused on "Comparative Sentiment Analysis," in which I will examine a variety of product
reviews to determine which products within the same category are more positively regarded by
customers. Businesses are continuously looking for methods to obtain a competitive edge as
competition rises across all industries. By examining the opinions expressed in user evaluations, one of
the most direct ways to measure consumer happiness and preferences is to do so. By harnessing the
power of NLP and sentiment analysis, I aim to shed light on the preferences and perceptions
that drive customer choices